Add custom CSS, layouts and includes for Octocopycat
    "Imitation is the sincerest form of flattery."

Octocopycat tries to mimic a lot of the GitHub layout, so that transitioning
from a user page to the user's GitHub profile will look quite seamless.

Furthermore, markdown content uses the same styles as the GitHub file browser,
so it looks quite familiar to those who have used GitHub before.

The CSS works both on mobile and on desktop site. On mobile, the header (with
the user's avatar and profile) is collapsed by default and becomes an icon at
the top.

The layouts can be customized through a site.octocopycat YAML fragment in the
site's _config.yml. (Documentation of the specific options is coming up.)

A navigation bar is populated from site.octocopycat.sitemap, which takes
objects with a title: and url: and uses them to populate the navbar.

The layout for posts (Blog) and pages will optionally use the page's title:
from the front matter and render it as an <h2>.